+++ main.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/MustWin/cmeter/cmd"
 	_ "github.com/MustWin/cmeter/cmd/agent"
-	_ "github.com/MustWin/cmeter/cmd/api"
 	"github.com/MustWin/cmeter/cmd/root"
 	_ "github.com/MustWin/cmeter/containers/embedded"
 	"github.com/MustWin/cmeter/context"