+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-### Added
+### Removed
+- the `api` command.
+- `mockapi` configuration section.
+- mock api server
 
-## [0.0.1] - 2016-09-16
+## [0.0.1] - 2016-09-19
 ### Added
 - select reporting driver in config file's `reporting` directive
 - select containers driver in config file's `containers` directive
@@ -15,6 +18,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `mock` reporting driver that just logs stuff out
 - basic pipeline implementation for processing messages
 - development config
+- mock api server
 - overridable and automatic build versioning support in Makefile
 - embedded cAdvisor containers driver
 - reporting driver abstraction